Refactor: reducer 분리
